Handling 400-500 error in separate handler now instead of via the metrics
