+++ src/app/interceptors.ts
@@ -1,6 +1,7 @@
 import {HTTP_INTERCEPTORS} from '@angular/common/http';
-import {HttpMetricsService} from './services/http-metrics.service';
 import {PERFECT_SCROLLBAR_CONFIG, PerfectScrollbarConfigInterface} from 'ngx-perfect-scrollbar';
+import {HttpMetricsService} from './services/http/interceptors/http-metrics.service';
+import {ErrorHandlerService} from './services/http/interceptors/error-handler.service';
 
 const DEFAULT_PERFECT_SCROLLBAR_CONFIG: PerfectScrollbarConfigInterface = {
     suppressScrollX: true
@@ -10,4 +11,5 @@ export const interceptorProviders =
     [
         {provide: PERFECT_SCROLLBAR_CONFIG, useValue: DEFAULT_PERFECT_SCROLLBAR_CONFIG},
         {provide: HTTP_INTERCEPTORS, useClass: HttpMetricsService, multi: true},
+        {provide: HTTP_INTERCEPTORS, useClass: ErrorHandlerService, multi: true},
     ];
