+++ src/app/services/http/interceptors/http-metrics.service.ts
@@ -1,10 +1,10 @@
 import {EventEmitter, Inject, Injectable, OnDestroy} from '@angular/core';
 import {APP_CONFIG, AppConfig} from 'src/app.config';
 import {NGXLogger} from 'ngx-logger';
-import {HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from '@angular/common/http';
-import {BehaviorSubject, Observable} from 'rxjs';
-import {catchError, finalize, tap} from 'rxjs/operators';
-import {BusyNotifierService} from './busy-notifier.service';
+import {HttpEvent, HttpHandler, HttpInterceptor, HttpRequest, HttpResponse} from '@angular/common/http';
+import {BehaviorSubject, Observable, throwError} from 'rxjs';
+import {catchError, finalize, map, tap} from 'rxjs/operators';
+import {BusyNotifierService} from '../../busy-notifier.service';
 
 @Injectable({
     providedIn: 'root'
@@ -24,10 +24,6 @@ export class HttpMetricsService implements HttpInterceptor {
     intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
         this.beginRequest();
         return next.handle(req).pipe(
-            catchError((err, caught) => {
-                this.notifier.failure(err);
-                return caught;
-            }),
             finalize(() => this.endRequest())
         );
     }
